change file structures
